Abstracted handler functions into controllers.
