+++ server/src/modules/appointments/infra/http/controllers/AppointmentsController.ts
@@ -1,0 +1,22 @@
+import { Request, Response } from 'express'
+import { container } from 'tsyringe'
+import { parseISO } from 'date-fns'
+
+import CreateAppointmentService from '@modules/appointments/services/CreateAppointmentService'
+
+export default class AppointmentsController {
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { provider_id, date } = request.body
+
+    const parsedDate = parseISO(date)
+
+    const createAppointment = container.resolve(CreateAppointmentService)
+
+    const appointment = await createAppointment.execute({
+      provider_id,
+      date: parsedDate,
+    })
+
+    return response.json(appointment)
+  }
+}
