+++ server/src/modules/appointments/infra/http/routes/appointments.routes.ts
@@ -1,32 +1,13 @@
 import { Router } from 'express'
-import { parseISO } from 'date-fns'
-import CreateAppointmentService from '@modules/appointments/services/CreateAppointmentService'
+
 import ensureAuthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated'
-import { container } from 'tsyringe'
+import AppointmentsController from '../controllers/AppointmentsController'
 
 const appointmentsRouter = Router()
+const appointmentsController = new AppointmentsController()
 
 appointmentsRouter.use(ensureAuthenticated)
 
-// appointmentsRouter.get('/', async (request, response) => {
-//   const appointments = await appointmentsRepository.find()
-
-//   return response.json(appointments)
-// })
-
-appointmentsRouter.post('/', async (request, response) => {
-  const { provider_id, date } = request.body
-
-  const parsedDate = parseISO(date)
-
-  const createAppointment = container.resolve(CreateAppointmentService)
-
-  const appointment = await createAppointment.execute({
-    provider_id,
-    date: parsedDate,
-  })
-
-  return response.json(appointment)
-})
+appointmentsRouter.post('/', appointmentsController.create)
 
 export default appointmentsRouter
