+++ server/src/modules/users/infra/http/controllers/UsersController.ts
@@ -1,0 +1,19 @@
+import { Request, Response } from 'express'
+import { container } from 'tsyringe'
+
+import CreateUserService from '@modules/users/services/CreateUserService'
+
+export default class UsersController {
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { name, email, password } = request.body
+
+    const createUser = container.resolve(CreateUserService)
+
+    const user = await createUser.execute({ name, email, password })
+
+    return response.status(201).json({
+      ...user,
+      password: undefined,
+    })
+  }
+}
