+++ server/src/modules/users/infra/http/routes/sessions.routes.ts
@@ -1,24 +1,10 @@
 import { Router } from 'express'
-import { container } from 'tsyringe'
 
-import AuthenticateUserService from '@modules/users/services/AuthenticateUserService'
+import SessionsController from '../controllers/SessionsController'
 
 const sessionsRouter = Router()
+const sessionsController = new SessionsController()
 
-sessionsRouter.post('/', async (request, response) => {
-  const { email, password } = request.body
-
-  const authenticateUser = container.resolve(AuthenticateUserService)
-
-  const { user, token } = await authenticateUser.execute({
-    email,
-    password,
-  })
-
-  return response.json({
-    user: { ...user, password: undefined },
-    token,
-  })
-})
+sessionsRouter.post('/', sessionsController.create)
 
 export default sessionsRouter
