+++ server/src/modules/users/infra/http/routes/users.routes.ts
@@ -1,46 +1,25 @@
 import { Router } from 'express'
-import { container } from 'tsyringe'
 import multer from 'multer'
 
 import uploadConfig from '@config/upload'
 
-import CreateUserService from '@modules/users/services/CreateUserService'
-import UpdateUserAvatarService from '@modules/users/services/UpdateUserAvatarService'
 import ensureAuthenticated from '../middlewares/ensureAuthenticated'
+import UsersController from '../controllers/UsersController'
+import UserAvatarController from '../controllers/UserAvatarController'
 
 const usersRouter = Router()
+const usersController = new UsersController()
+const userAvatarController = new UserAvatarController()
 
 const upload = multer(uploadConfig)
 
-usersRouter.post('/', async (request, response) => {
-  const { name, email, password } = request.body
-
-  const createUser = container.resolve(CreateUserService)
-
-  const user = await createUser.execute({ name, email, password })
-
-  return response.status(201).json({
-    ...user,
-    password: undefined,
-  })
-})
+usersRouter.post('/', usersController.create)
 
 usersRouter.patch(
   '/avatar',
   ensureAuthenticated,
   upload.single('avatar'),
-  async (request, response) => {
-    const updateUserAvatar = container.resolve(UpdateUserAvatarService)
-
-    const user = await updateUserAvatar.execute({
-      user_id: request.user.id,
-      avatarFilename: request.file.filename,
-    })
-
-    delete user.password
-
-    return response.json(user)
-  },
+  userAvatarController.update,
 )
 
 export default usersRouter
